Time: 2 ms (99.42%) | Memory: 42.3 MB (12.60%) - LeetSync

+++ 1064-smallest-integer-divisible-by-k/smallest-integer-divisible-by-k.java
@@ -1,7 +1,7 @@
 class Solution {
     public int smallestRepunitDivByK(int k) {
         if(k%2==0||k%5==0){return -1;}
-        int n=1;
+        long n=1;
         int ans=1;
         boolean t=true;
         while(true){
